# openapi.yaml
openapi: 3.0.0
info:
  title: TiansuoVM API
  description: TiansuoVM平台API文档
  version: 1.0.0
servers:
  - url: /api/v1
    description: TiansuoVM API基础路径
...
paths:
...
  /admin/users/{uid}:
    put:
      summary: 更新用户信息(管理员)
      tags:
        - 管理员功能
      operationId: updateUser
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
    delete:
      summary: 删除用户(管理员)
      tags:
        - 管理员功能
      operationId: deleteUser
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] 